COMMERCE-3777 - portal-instances-api - Rename ”initalizer” to “initializer”
