+++ modules/apps/portal-instances/portal-instances-api/src/main/java/com/liferay/portal/instances/initializer/PortalInstanceInitializerRegistry.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.instances.initalizer;
+package com.liferay.portal.instances.initializer;
 
-import com.liferay.portal.instances.exception.InitializationException;
+import java.util.List;
 
 import org.osgi.annotation.versioning.ProviderType;
 
@@ -22,13 +22,13 @@ import org.osgi.annotation.versioning.ProviderType;
  * @author Ivica Cardic
  */
 @ProviderType
-public interface PortalInstanceInitializer {
+public interface PortalInstanceInitializerRegistry {
 
-	public String getKey();
+	public PortalInstanceInitializer getPortalInstanceInitializer(String key);
 
-	public void initialize(String webId, String virtualHostname, String mx)
-		throws InitializationException;
+	public List<PortalInstanceInitializer> getPortalInstanceInitializers();
 
-	public boolean isActive();
+	public List<PortalInstanceInitializer> getPortalInstanceInitializers(
+		boolean activeOnly);
 
 }
